Level of Detail for Nodes and TrainrunSection is working (#107)
An implementation of different levels of detail (Levels) has been carried out, including Full, Level3, Level2, Level1, and Level0. For each level of detail, a corresponding "make object" function has been developed to create the respective objects.

Additionally, a "Level of Detail" service has been implemented to manage and control the different levels of detail. This service ensures that the appropriate level of detail is loaded and displayed according to the requirements.

In connection with this, methods for visibility management of viewports (Viewport Culling) have been moved to a separate "Viewport Culling Service." This service is responsible for hiding or optimizing objects outside the visible area to improve performance.

Overall, work has been done to optimize the rendering and performance, enabling efficient and adaptable representation of objects.