feat: publish package on NPM
Add two new CI tasks: one to build the standalone variant on push,
and another one to publish an NPM package on release.

The latter requires setting up a `NPM_TOKEN` secret in the GitHub
repository settings.

All dependencies are removed before publication because the
published package is completely standalone: all of the dependencies
are bundled in `dist/`.

Add the "repository" field in package.json: it's a requirement for
the npm publish --provenance flag.