refactor: ensure consistent source → target trainrunSections chain
This commit intends to ensure that the trainruns are all [source] → [target] consistent.

Before this commit, we could have trainruns like [[source] → [target][target] → [source][target] → [target][source]...], which is a pain to work with because the trainrunSections are not ordered, so the order has to be deducted from the [source] → [target] chain.

It is an enabler for one-way trainruns, since we need to have ordered trainrunSections.

The current implementation has been stolen to @emersion, and is not compatible with actual cyclic trainruns. It needs to be rethought.

Signed-off-by: Louis Greiner <greiner.louis@gmail.com>